Menu problem 50% resolved
Reminder for Monday: Canvas elements makes it impossible to use the menu properly. When clicking the collapse button and then extending the menu again: the main content area will break and it will be put under the menu. This happens due to the canvas elements that includes the chart. This should have a re-work on monday.